refactor: move findRuntimeNodeVersion / readSnapshotRuntimePin to @pnpm/deps.graph-hasher
`@pnpm/engine.runtime.system-node-version`'s purpose is probing the
*host* Node — `getSystemNodeVersion`, `engineName`. The two lockfile-
shape parsers (`findRuntimeNodeVersion`, `readSnapshotRuntimePin`)
were a poor fit there and are now exported from the package that
actually composes the engine string: `@pnpm/deps.graph-hasher`. Every
caller already depended on graph-hasher, so no new deps are introduced;
six packages drop the now-unused dependency on system-node-version.

The pre-existing `.changeset/gvs-engine-name-shell-node.md` (#11689,
not yet released) is updated to reflect the new home for
`findRuntimeNodeVersion`.
