+++ building/after-install/tsconfig.json
@@ -40,9 +40,6 @@
       "path": "../../deps/path"
     },
     {
-      "path": "../../engine/runtime/system-node-version"
-    },
-    {
       "path": "../../exec/lifecycle"
     },
     {
