+++ building/during-install/tsconfig.json
@@ -37,9 +37,6 @@
       "path": "../../deps/path"
     },
     {
-      "path": "../../engine/runtime/system-node-version"
-    },
-    {
       "path": "../../exec/lifecycle"
     },
     {
