+++ deps/graph-builder/tsconfig.json
@@ -25,9 +25,6 @@
       "path": "../../core/types"
     },
     {
-      "path": "../../engine/runtime/system-node-version"
-    },
-    {
       "path": "../../hooks/types"
     },
     {
