+++ engine/runtime/system-node-version/src/index.ts
@@ -27,8 +27,9 @@ export const getSystemNodeVersion = mem(getSystemNodeVersionNonCached)
  *
  * 1. `nodeVersion` argument when provided. Callers use this to thread
  *    a project-pinned runtime (`engines.runtime` / `devEngines.runtime`)
- *    through to the hash — see {@link findRuntimeNodeVersion} for the
- *    helper that extracts the value from a lockfile.
+ *    through to the hash — see `findRuntimeNodeVersion` /
+ *    `readSnapshotRuntimePin` in `@pnpm/deps.path` for the helpers
+ *    that extract the value from a lockfile or graph node.
  * 2. {@link getSystemNodeVersion} — the `node` on the user's `PATH`,
  *    or `process.version` when not SEA-bundled.
  * 3. `process.version` as a last-resort fallback when the host has
@@ -51,73 +52,4 @@ export function engineName (nodeVersion?: string): string {
   const stripped = version.startsWith('v') ? version.slice(1) : version
   const major = stripped.split('.')[0]
   return `${process.platform};${process.arch};node${major}`
-}
-
-/**
- * Strip the `node@runtime:` prefix and any peer-context suffix `(...)`
- * from a single snapshot key, returning the bare Node version (e.g.
- * `"22.11.0"`) — or `undefined` if the key isn't a Node runtime pin.
- *
- * Peer-suffixed (`node@runtime:22.11.0(node@22.11.0)`) and bare
- * (`node@runtime:22.11.0`) forms must reduce to the same answer; the
- * pacquet side relies on the same rule for GVS-hash parity.
- */
-function extractRuntimeNodeVersion (snapshotKey: string): string | undefined {
-  const prefix = 'node@runtime:'
-  if (!snapshotKey.startsWith(prefix)) return undefined
-  const versionWithPeers = snapshotKey.slice(prefix.length)
-  const parenAt = versionWithPeers.indexOf('(')
-  return parenAt === -1 ? versionWithPeers : versionWithPeers.slice(0, parenAt)
-}
-
-/**
- * Scan an iterable of lockfile snapshot keys for the resolved
- * `engines.runtime` / `devEngines.runtime` Node version and return
- * its bare version string (e.g. `"22.11.0"`), or `undefined` when
- * no snapshot pins a runtime.
- *
- * Pnpm's runtime resolver writes the pinned Node into the lockfile as
- * a snapshot with key `node@runtime:<version>[(<peers>)]`
- * (see [`engine/runtime/node-resolver/src/index.ts`](https://github.com/pnpm/pnpm/blob/29a42efc3b/engine/runtime/node-resolver/src/index.ts)).
- * The first such key found is treated as authoritative. This is fine
- * as an install-wide fallback (project-pin in the typical case), but
- * snapshots that pin their own Node still need
- * {@link readSnapshotRuntimePin} to get a per-snapshot result.
- *
- * Callers typically pass `Object.keys(lockfile.packages ?? {})` — the
- * in-memory `LockfileObject` merges the on-disk `packages:` and
- * `snapshots:` sections under a single `packages` field, so its keys
- * include every snapshot key the install will hash.
- */
-export function findRuntimeNodeVersion (snapshotKeys: Iterable<string>): string | undefined {
-  for (const key of snapshotKeys) {
-    const version = extractRuntimeNodeVersion(key)
-    if (version != null) return version
-  }
-  return undefined
-}
-
-/**
- * Read a single graph node's own `engines.runtime` Node pin from its
- * `children` map. The resolver desugars `engines.runtime` declared on
- * a dependency's manifest into `dependencies.node: 'runtime:<version>'`
- * (see [`installing/deps-resolver/src/resolveDependencies.ts`](https://github.com/pnpm/pnpm/blob/29a42efc3b/installing/deps-resolver/src/resolveDependencies.ts)),
- * which then becomes a `children.node` entry pointing at the
- * `node@runtime:<version>[(peers)]` snapshot key.
- *
- * Returns the bare version (e.g. `"22.11.0"`) when this snapshot pins
- * its own Node — or `undefined` when it doesn't and the caller should
- * fall back to the install-wide pin / host probe.
- *
- * Per-snapshot resolution matters because the bin linker routes
- * lifecycle-script spawns for a pinning package through *that
- * package's* downloaded Node — anchoring the snapshot's GVS engine
- * hash to an install-wide value would mis-key the side-effects cache
- * for cross-pinning installs.
- */
-export function readSnapshotRuntimePin (
-  children: Record<string, string> | undefined
-): string | undefined {
-  const ref = children?.node
-  return ref != null ? extractRuntimeNodeVersion(ref) : undefined
 }
