+++ installing/deps-installer/tsconfig.json
@@ -91,9 +91,6 @@
       "path": "../../deps/path"
     },
     {
-      "path": "../../engine/runtime/system-node-version"
-    },
-    {
       "path": "../../exec/lifecycle"
     },
     {
