+++ installing/deps-restorer/package.json
@@ -48,7 +48,6 @@
     "@pnpm/deps.graph-builder": "workspace:*",
     "@pnpm/deps.graph-hasher": "workspace:*",
     "@pnpm/deps.path": "workspace:*",
-    "@pnpm/engine.runtime.system-node-version": "workspace:*",
     "@pnpm/error": "workspace:*",
     "@pnpm/exec.lifecycle": "workspace:*",
     "@pnpm/fs.symlink-dependency": "workspace:*",
